Separated context and block operations in persistence
